# doc/api.yaml
openapi: 3.0.0
paths:
...
  /artist/{id}:
    get:
      operationId: ArtistController_findOne
      summary: get single artist by id
      parameters:
        - name: id
          required: true
          in: path
          description: Id of artist
          schema:
            type: string
      responses:
        '200':
          description: Artist successfully getted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Artist id is invalid (not UUID)
        '404':
          description: Artist with given id not found
      tags:
        - Artist
      security:
        - bearer: []
    put:
      operationId: ArtistController_update
      summary: update artist info
      parameters:
        - name: id
          required: true
          in: path
          description: Id of artist
          schema:
            type: string
      requestBody:
        required: true
        description: Data for update album
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: Artist successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Artist id is invalid (not UUID)
        '404':
          description: Artist with given id not found
      tags:
        - Artist
      security:
        - bearer: []
    delete:
      operationId: ArtistController_remove
      summary: delete artist
      parameters:
        - name: id
          required: true
          in: path
          description: Id of artist
          schema:
            type: string
      responses:
        '204':
          description: Artist successfully deleted
        '400':
          description: Artist id is invalid (not UUID)
        '404':
          description: Artist with given id not found
      tags:
        - Artist
      security:
        - bearer: []
...
  /album/{id}:
    get:
      operationId: AlbumController_findOne
      summary: get single album by id
      parameters:
        - name: id
          required: true
          in: path
          description: Id of album
          schema:
            type: string
      responses:
        '200':
          description: Album successfully getted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Album id is invalid (not UUID)
        '404':
          description: Album with given id not found
      tags:
        - Album
      security:
        - bearer: []
    put:
      operationId: AlbumController_update
      summary: update album info
      parameters:
        - name: id
          required: true
          in: path
          description: Id of album
          schema:
            type: string
      requestBody:
        required: true
        description: Data for update album
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: Album successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Album id is invalid (not UUID)
        '404':
          description: Album with given id not found
      tags:
        - Album
      security:
        - bearer: []
    delete:
      operationId: AlbumController_remove
      summary: delete album
      parameters:
        - name: id
          required: true
          in: path
          description: Id of album
          schema:
            type: string
      responses:
        '204':
          description: Album successfully deleted
        '400':
          description: Album id is invalid (not UUID)
        '404':
          description: Album with given id not found
      tags:
        - Album
      security:
        - bearer: []
  /track:
    get:
      operationId: TrackController_findAll
      summary: get all tracks
      parameters: []
      responses:
        '200':
          description: Tracks successfully getted
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
      tags:
        - Track
      security:
        - bearer: []
    post:
      operationId: TrackController_create
      summary: create new track
      parameters: []
      requestBody:
        required: true
        description: Data for new track
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: Track successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Request body does not contain required fields
      tags:
        - Track
      security:
        - bearer: []
  /track/{id}:
    get:
      operationId: TrackController_findOne
      summary: get single track by id
      parameters:
        - name: id
          required: true
          in: path
          description: Id of track
          schema:
            type: string
      responses:
        '200':
          description: Track successfully getted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Track id is invalid (not UUID)
        '404':
          description: Track with given id not found
      tags:
        - Track
      security:
        - bearer: []
    put:
      operationId: TrackController_update
      summary: update album track
      parameters:
        - name: id
          required: true
          in: path
          description: Id of track
          schema:
            type: string
      requestBody:
        required: true
        description: Data for update album
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: Track successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Track id is invalid (not UUID)
        '404':
          description: Track with given id not found
      tags:
        - Track
      security:
        - bearer: []
    delete:
      operationId: TrackController_remove
      summary: delete track
      parameters:
        - name: id
          required: true
          in: path
          description: Id of track
          schema:
            type: string
      responses:
        '204':
          description: Track successfully deleted
        '400':
          description: Track id is invalid (not UUID)
        '404':
          description: Track with given id not found
      tags:
        - Track
      security:
        - bearer: []
  /favs:
    get:
      operationId: FavoritesController_findAll
      summary: get all favorites
      parameters: []
      responses:
        '200':
          description: Favorites successfully getted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorite'
      tags:
        - Favorites
      security:
        - bearer: []
  /favs/track/{id}:
    post:
      operationId: FavoritesController_addTrackToFavorites
      summary: add track to the favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Id of track
          schema:
            type: string
      responses:
        '201':
          description: Track successfully added
        '400':
          description: Track id is invalid (not UUID).
        '422':
          description: Track with given id does not exist
      tags:
        - Favorites
      security:
        - bearer: []
    delete:
      operationId: FavoritesController_removeTrackFromFavorites
      summary: delete track from favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Id of track
          schema:
            type: string
      responses:
        '204':
          description: Track successfully deleted
        '400':
          description: Track id is invalid (not UUID)
        '404':
          description: Track with given id does not found
      tags:
        - Favorites
      security:
        - bearer: []
  /favs/album/{id}:
    post:
      operationId: FavoritesController_addAlbumToFavorites
      summary: add album to the favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Id of album
          schema:
            type: string
      responses:
        '201':
          description: Album successfully added
        '400':
          description: Album id is invalid (not UUID).
        '422':
          description: Album with given id does not exist
      tags:
        - Favorites
      security:
        - bearer: []
    delete:
      operationId: FavoritesController_removeAlbumFromFavorites
      summary: delete album from favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Id of album
          schema:
            type: string
      responses:
        '204':
          description: Album successfully deleted
        '400':
          description: Album id is invalid (not UUID)
        '404':
          description: Album with given id does not found
      tags:
        - Favorites
      security:
        - bearer: []
  /favs/artist/{id}:
    post:
      operationId: FavoritesController_addArtistToFavorites
      summary: add artist to the favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Id of artist
          schema:
            type: string
      responses:
        '201':
          description: Artist successfully added
        '400':
          description: Artist id is invalid (not UUID).
        '422':
          description: Artist with given id does not exist
      tags:
        - Favorites
      security:
        - bearer: []
    delete:
      operationId: FavoritesController_removeArtistFromFavorites
      summary: delete artist from favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Id of artist
          schema:
            type: string
      responses:
        '204':
          description: Artist successfully deleted
        '400':
          description: Artist id is invalid (not UUID)
        '404':
          description: Artist with given id does not found
      tags:
        - Favorites
      security:
        - bearer: []
...
info:
  title: Home Library Service
  description: 'Task #7: Implementation Home Library Service: Part 3'
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    Artist:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        grammy:
          type: boolean
      required:
        - id
        - name
        - grammy
...
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
      required:
        - name
        - grammy
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          format: uuid
          nullable: true
      required:
        - id
        - name
        - year
        - artistId
...
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          format: uuid
          nullable: true
      required:
        - name
        - year
        - artistId
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        artistId:
          type: string
          format: uuid
          nullable: true
        albumId:
          type: string
          format: uuid
          nullable: true
        duration:
          type: number
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
        artistId:
          type: string
          format: uuid
          nullable: true
        albumId:
          type: string
          format: uuid
          nullable: true
        duration:
          type: number
      required:
        - name
        - artistId
        - albumId
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
        artistId:
          type: string
          format: uuid
          nullable: true
        albumId:
          type: string
          format: uuid
          nullable: true
        duration:
          type: number
      required:
        - name
        - artistId
        - albumId
        - duration
    Favorite:
      type: object
      properties:
        artists:
          type: array
          items:
            $ref: '#/components/schemas/Artist'
        albums:
          type: array
          items:
            $ref: '#/components/schemas/Album'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
      required:
        - artists
        - albums
        - tracks
...